# Conflicts:
#	src/Showcase/src/Smart.FA.Catalog.Showcase.Web/Pages/Training/TrainingList/TrainingList.cshtml.cs
#	src/Showcase/src/Smart.FA.Catalog.Showcase.Web/Pages/Training/TrainingList/TrainingListViewModel.cs
#	src/UserAdmin/src/Smart.FA.Catalog.Core/Domain/Trainer/Trainer.cs
#	src/UserAdmin/src/Smart.FA.Catalog.Infrastructure/Migrations/ContextModelSnapshot.cs

+++ src/Showcase/src/Smart.FA.Catalog.Showcase.Domain/Models/TrainingList.cs
@@ -11,6 +11,6 @@ public partial class TrainingList
     public string TrainerFirstName { get; set; } = null!;
     public string TrainerLastName { get; set; } = null!;
     public int Topic { get; set; }
-    public short Status { get; set; }
+    public int Status { get; set; }
     public int TrainerId { get; set; }
 }
